Feat: User permission validation for group participation methods

+++ src/main/java/br/edu/ifsp/scl/pipegene/usecases/group/GroupCRUDimpl.java
@@ -8,6 +8,7 @@ import br.edu.ifsp.scl.pipegene.usecases.account.gateway.UserApplicationDAO;
 import br.edu.ifsp.scl.pipegene.usecases.account.model.ApplicationUser;
 import br.edu.ifsp.scl.pipegene.usecases.group.gateway.GroupDAO;
 import br.edu.ifsp.scl.pipegene.web.exception.ResourceNotFoundException;
+import org.springframework.dao.PermissionDeniedDataAccessException;
 import org.springframework.stereotype.Service;
 
 import java.sql.Timestamp;
@@ -39,6 +40,8 @@ public class GroupCRUDimpl implements GroupCRUD{
 
     @Override
     public GroupParticipation addToGroup(UUID groupId, String username) {
+        GroupParticipation groupParticipation = null;
+
         Optional<ApplicationUser> userOptional = userApplicationDAO.findUserByUsername(username);
         if (userOptional.isEmpty())
             throw new ResourceNotFoundException("Not found user with username: " + username);
@@ -47,7 +50,20 @@ public class GroupCRUDimpl implements GroupCRUD{
         if (groupOptional.isEmpty())
             throw new ResourceNotFoundException("Not found group with id: " + groupId);
 
-        var groupParticipation = GroupParticipation.createWithAllFields(UUID.randomUUID(), groupOptional.get(), userOptional.get().getId(), GroupParticipationStatusEnum.PENDING, authentication.getUserAuthenticatedId(), Timestamp.from(now()));
+        Optional<GroupParticipation> groupParticipationOptional = groupDAO.findGroupParticipationByGroupIdAndReceiverId(groupId, userOptional.get().getId());
+        if (groupParticipationOptional.isPresent()){
+            groupParticipation = groupParticipationOptional.get();
+            if (groupParticipation.getStatus().equals(GroupParticipationStatusEnum.ACCEPTED))
+                throw new PermissionDeniedDataAccessException("User already in group", null);
+            if (groupParticipation.getStatus().equals(GroupParticipationStatusEnum.PENDING))
+                throw new PermissionDeniedDataAccessException("User already invited to the group", null);
+            groupParticipation.setStatus(GroupParticipationStatusEnum.PENDING);
+            groupParticipation.setCreatedDate(Timestamp.from(now()));
+            groupDAO.updateGroupParticipation(groupParticipation);
+            return groupParticipation;
+        }
+
+        groupParticipation = GroupParticipation.createWithAllFields(UUID.randomUUID(), groupOptional.get(), userOptional.get().getId(), GroupParticipationStatusEnum.PENDING, authentication.getUserAuthenticatedId(), Timestamp.from(now()));
         return groupDAO.saveGroupParticipation(groupParticipation);
     }
 
@@ -72,10 +88,15 @@ public class GroupCRUDimpl implements GroupCRUD{
     @Override
     public GroupParticipation deleteGroupParticipation(UUID id){
         Optional<GroupParticipation> groupParticipationOptional = groupDAO.findGroupParticipationById(id);
+        UUID userId = authentication.getUserAuthenticatedId();
         if (groupParticipationOptional.isEmpty())
             throw new ResourceNotFoundException("Not found group participation with id: " + id);
 
         var groupParticipation = groupParticipationOptional.get();
+
+        if (userId != groupParticipation.getSubmitterId())
+            throw new PermissionDeniedDataAccessException("You don't have permission to deny this group participation", null);
+
         return groupDAO.deleteGroupParticipation(groupParticipation.getId());
     }
 
@@ -85,6 +106,11 @@ public class GroupCRUDimpl implements GroupCRUD{
             throw new ResourceNotFoundException("Not found group participation with id: " + groupParticipationId);
 
         var groupParticipation = groupParticipationOptional.get();
+
+        UUID userId = authentication.getUserAuthenticatedId();
+        if (!userId.equals(groupParticipation.getReceiverId()))
+            throw new PermissionDeniedDataAccessException("You don't have permission to deny this group participation", null);
+
         return groupParticipation;
     }
